Most files regenerated successfully
This commit closes #9 by adding the new girgen/file.go FileGenerator,
which is completely refactored from the ground up to be much smaller,
cleaner and side effects-free.

This commit closes #8 as well by introducing the SourceNamespace and
CurrentNamespace API inside typeconv.Converter, which allows the
converter to pick the right namespace for resolving and generating.