Saner enum type, record pointers; union support
This commit changes most record return values to return a pointer
instead of value, to both reflect how it'll always pass-by-reference and
how most record methods need a pointer receiver.

This commit also changes the underlying types for enums and bitfields to
be C.int and C.uint, respectively, to allow for direct casting of
enums/bitfields lists.

The record pointer change is expected to be followed up with a similar
change for classes. This will make signal generation slightly more
accurate as well, since both of those types are always passed by
pointers.

This commit closes issue #25.