Implement glib.GeneratedClosure
This commit implements a glib.GeneratedClosure wrapper for closures that
are passed into (*glib.Object).Connect(). This introduces a side effect
that the first argument check is always assumed to be true, that is,
generated ConnectX methods will never have the first argument be the
receiver value.

A slight regression of this commit is closure.NewFuncStack not
type-asserting that the value is a function anymore, since
GeneratedClosure is a wrapper struct which fails the check.