Refactor GTK documentation parser (#140)
* Migrate from github.com/pkg/errors.Wrap to stdlib %w

* gir/girgen/cmt: Comment which things in "go/doc" are deprecated

* gir/girgen/cmt: Tidy up

This is tidying up in prep for me making real changes.  None of the
changes here should be functional changes.

* gir-generate/gendata: Drop the unused Package type

* Rethink pkgconfig

The main purpose of this is to let me get other variables out of
pkgconfig, which is functionality that I will use in later work.

* gir/girgen/cmt: Don't print param/ret labels if we don't print the list

* gir/girgen/cmt: Use the new go/doc/comment package for formatting

As you can see, I got its formatting to exactly match the old formatting.
This allows us to be sure that this doesn't introduce any regressions.
We can make improvements to the formatting in subsequent commits.

* gir/girgen/cmt: Drop back-compat whitespace hacks

I added these in the previous commit so that the formatting would match
exactly.  Now that we've verified that there are no regressions, let's
drop them.

This brings the formatting in-line with `gofmt`, so go ahead and have the
"lint" CI job check the generated files too.